Step 3: Extract MessageIndexer out of ElasticSearchMessageSearchIndex
