+++ mailbox/elasticsearch/src/test/java/org/apache/james/mailbox/elasticsearch/ElasticSearchIntegrationTest.java
@@ -37,12 +37,9 @@ import org.apache.james.backends.es.utils.TestingClientProvider;
 import org.apache.james.mailbox.MessageManager;
 import org.apache.james.mailbox.acl.SimpleGroupMembershipResolver;
 import org.apache.james.mailbox.elasticsearch.events.ElasticSearchListeningMessageSearchIndex;
+import org.apache.james.mailbox.elasticsearch.events.ElasticSearchMessageIndexer;
 import org.apache.james.mailbox.elasticsearch.json.MessageToElasticSearchJson;
-import org.apache.james.mailbox.elasticsearch.json.MessageToElasticSearchJsonV1;
-import org.apache.james.mailbox.elasticsearch.query.CriterionConverter;
-import org.apache.james.mailbox.elasticsearch.query.QueryConverter;
 import org.apache.james.mailbox.elasticsearch.search.ElasticSearchSearcher;
-import org.apache.james.mailbox.elasticsearch.search.ElasticSearchSearcherV1;
 import org.apache.james.mailbox.extractor.TextExtractor;
 import org.apache.james.mailbox.inmemory.InMemoryId;
 import org.apache.james.mailbox.inmemory.manager.InMemoryIntegrationResources;
@@ -122,9 +119,7 @@ public abstract class ElasticSearchIntegrationTest extends AbstractMessageSearch
         storeMailboxManager = new InMemoryIntegrationResources()
             .createMailboxManager(new SimpleGroupMembershipResolver());
 
-
-        ElasticSearchListeningMessageSearchIndex elasticSearchListeningMessageSearchIndex = new ElasticSearchListeningMessageSearchIndex(
-            storeMailboxManager.getMapperFactory(),
+        ElasticSearchMessageIndexer indexerDelegate = new ElasticSearchMessageIndexer(
             new ElasticSearchIndexer(client,
                 new DeleteByQueryPerformer(client,
                     Executors.newSingleThreadExecutor(),
@@ -133,8 +128,11 @@ public abstract class ElasticSearchIntegrationTest extends AbstractMessageSearch
                     MailboxElasticSearchConstants.MESSAGE_TYPE),
                 MailboxElasticSearchConstants.DEFAULT_MAILBOX_WRITE_ALIAS,
                 MailboxElasticSearchConstants.MESSAGE_TYPE),
-            provideSearcher(client, new InMemoryId.Factory(), storeMailboxManager.getMessageIdFactory()),
             provideMessageToElasticSearchJson(textExtractor,  ZoneId.of("Europe/Paris")));
+
+        ElasticSearchListeningMessageSearchIndex elasticSearchListeningMessageSearchIndex = new ElasticSearchListeningMessageSearchIndex(storeMailboxManager.getMapperFactory(),
+            provideSearcher(client, new InMemoryId.Factory(), storeMailboxManager.getMessageIdFactory()),
+            indexerDelegate);
 
         messageIdManager = new StoreMessageIdManager(
             storeMailboxManager,
